+++ libraries/I2C_LCD/I2C_LCD.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: I2C_LCD.cpp
 //  AUTHOR: Rob.Tillaart@gmail.com
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 //    DATE: 2023-12-16
 // PUPROSE: Arduino library for I2C_LCD
 //     URL: https://github.com/RobTillaart/I2C_LCD
@@ -9,7 +9,14 @@
 
 #include "I2C_LCD.h"
 
+//  40 us is a save value
+const uint8_t I2C_LCD_CHAR_DELAY = 40;
 
+
+///////////////////////////////////////////////////////
+//
+//  DO NOT CHANGE BELOW THIS LINE
+//
 //  keep defines compatible / recognizable
 //  the zero valued defines are not used.
 #define I2C_LCD_CLEARDISPLAY        0x01
@@ -36,6 +43,9 @@
 #define I2C_LCD_5x10DOTS            0x04
 
 
+
+
+
 I2C_LCD::I2C_LCD(uint8_t address, TwoWire * wire)
 {
   _address = address;
@@ -46,7 +56,7 @@ I2C_LCD::I2C_LCD(uint8_t address, TwoWire * wire)
 
 void I2C_LCD::config (uint8_t address, uint8_t enable, uint8_t readWrite, uint8_t registerSelect,
                       uint8_t data4, uint8_t data5, uint8_t data6, uint8_t data7,
-                      uint8_t backLight, uint8_t policy)  
+                      uint8_t backLight, uint8_t polarity)  
 {
   if (_address != address) return;  //  compatible?
   _enable         = ( 1 << enable);
@@ -57,7 +67,7 @@ void I2C_LCD::config (uint8_t address, uint8_t enable, uint8_t readWrite, uint8_
   _dataPin[2]     = ( 1 << data6);
   _dataPin[3]     = ( 1 << data7);
   _backLightPin   = ( 1 << backLight);
-  _backLightPol   = policy;
+  _backLightPol   = polarity;
 
   _pinsInOrder = ((data4 < data5) && (data5 < data6) && (data6 < data7));
 }
@@ -68,15 +78,14 @@ void I2C_LCD::begin(uint8_t cols, uint8_t rows)
   _cols = cols;
   _rows = rows;
 
-  //  ALL LOW.
+  //  ALL LINES LOW.
   _wire->beginTransmission(_address);
-  _wire->write(0x00);
   _wire->write(0x00);
   _wire->endTransmission();
 
   //  Figure 24 for procedure on 4-bit initialization
   //  wait for more than 15 ms
-  delay(60);  //  TODO optimize
+  delay(100);  //  no need to optimize as this is called only once.
 
   //  Force 4 bit mode
   write4bits(0x03);
@@ -109,17 +118,18 @@ bool I2C_LCD::isConnected()
 //
 //  BACKLIGHT
 //
-void I2C_LCD::setBacklightPin(uint8_t pin, uint8_t policy)
+void I2C_LCD::setBacklightPin(uint8_t pin, uint8_t polarity)
 {
   _backLightPin = ( 1 << pin);
-  _backLightPol = policy;
+  _backLightPol = polarity;
 }
 
 
 void I2C_LCD::setBacklight(bool on)
 {
-  _backLight = on;
-  display();
+  _backLight = (on == _backLightPol);
+  if (_backLight) display();
+  else noDisplay();
 }
 
 
@@ -275,10 +285,7 @@ void I2C_LCD::createChar(uint8_t index, uint8_t * charmap)
 }
 
 
-//////////////////////////////////////////////////////////
-//
-//  PRIVATE
-//
+
 size_t I2C_LCD::write(uint8_t c)
 {
   size_t n = 0;
@@ -303,57 +310,46 @@ size_t I2C_LCD::write(uint8_t c)
 };
 
 
+size_t I2C_LCD::center(uint8_t row, const char * message)
+{
+  uint8_t len = strlen(message) + 1;
+  setCursor(10 - len/2, row);
+  return print(message);
+}
 
-//  TODO merge these two
-//  optimize 95% identical.
 
-//  TODO merge these two
-//  optimize 95% identical.
+size_t I2C_LCD::right(uint8_t col, uint8_t row, const char * message)
+{
+  uint8_t len = strlen(message);
+  setCursor(col - len, row);
+  return print(message);
+}
+
+
+//////////////////////////////////////////////////////////
+//
+//  PRIVATE
+//
 void I2C_LCD::sendCommand(uint8_t value)
 {
-  uint8_t MSN = 0;
-  if (_backLight)  MSN |= _backLightPin;
-  uint8_t LSN = MSN;
-
-  //  pins are in the right order optimization.
-  if (_pinsInOrder)
-  {
-    MSN |= value & 0xF0;
-    LSN |= value << 4;
-  }
-  else
-  {
-    uint8_t tmp = value >> 4;
-    for ( uint8_t i = 0; i < 4; i++ )
-    {
-      if ( tmp & 0x01 ) MSN |= _dataPin[i];
-      tmp >>= 1;
-    }
-    tmp = value & 0x0F;
-    for ( uint8_t i = 0; i < 4; i++ )
-    {
-      if ( tmp & 0x01 ) LSN |= _dataPin[i];
-      tmp >>= 1;
-    }
-  }
-
-  _wire->beginTransmission(_address);
-  _wire->write(MSN | _enable);
-  _wire->write(MSN);
-  _wire->write(LSN | _enable);
-  _wire->write(LSN);
-  _wire->endTransmission();
-  delayMicroseconds(40);
+  send(value, false);
 }
 
 
 void I2C_LCD::sendData(uint8_t value)
 {
-  uint8_t MSN = _registerSelect;
+  send(value, true);
+}
+
+
+void I2C_LCD::send(uint8_t value, bool dataFlag)
+{
+  //  calculate both most and least significant nibble
+  uint8_t MSN = 0;
+  if (dataFlag)   MSN = _registerSelect;
   if (_backLight) MSN |= _backLightPin;
   uint8_t LSN = MSN;
 
-  //  if pins are in the right order speed up.
   if (_pinsInOrder)
   {
     MSN |= value & 0xF0;
@@ -361,17 +357,15 @@ void I2C_LCD::sendData(uint8_t value)
   }
   else
   {
-    uint8_t tmp = value >> 4;
     for ( uint8_t i = 0; i < 4; i++ )
     {
-      if ( tmp & 0x01 ) MSN |= _dataPin[i];
-      tmp >>= 1;
+      if ( value & 0x01 ) LSN |= _dataPin[i];
+      value >>= 1;
     }
-    tmp = value & 0x0F;
     for ( uint8_t i = 0; i < 4; i++ )
     {
-      if ( tmp & 0x01 ) LSN |= _dataPin[i];
-      tmp >>= 1;
+      if ( value & 0x01 ) MSN |= _dataPin[i];
+      value >>= 1;
     }
   }
 
@@ -381,15 +375,14 @@ void I2C_LCD::sendData(uint8_t value)
   _wire->write(LSN | _enable);
   _wire->write(LSN);
   _wire->endTransmission();
-  delayMicroseconds(40);
+  delayMicroseconds(I2C_LCD_CHAR_DELAY);
 }
 
 
-
-//  really needed for setup
+//  needed for setup
 void I2C_LCD::write4bits(uint8_t value) 
 {
-  uint8_t cmd = _backLight;
+  uint8_t cmd = 0;
 
   for ( uint8_t i = 0; i < 4; i++ )
   {