+++ libraries/I2C_LCD/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2023-12-21
+- implement polarity backlight (and removed policy)
+- add **size_t center(uint8_t row, const char \* message)**
+- add **size_t right(uint8_t col, uint8_t row, const char \* message)**
+- merged low level code reducing footprint.
+- optimized low level code when pins are not in ascending order (minor).
+- add examples
+- fix keywords.txt
+- update readme.md
+
+
 ## [0.1.1] - 2023-12-18
 - add blink() and noBlink()
 - add cursor() and noCursor()